+++ pkg/interop/errors.go
@@ -19,35 +19,54 @@ import (
 )
 
 var (
-	errNoPublicTLSAddress  = errors.DefineFailedPrecondition("no_public_tls_address", "no public TLS address configured for interop")
-	errUnknownMACVersion   = errors.DefineInvalidArgument("unknown_mac_version", "unknown MAC version")
-	errInvalidLength       = errors.DefineInvalidArgument("invalid_length", "invalid length")
-	errInvalidRequestType  = errors.DefineInvalidArgument("invalid_request_type", "invalid request type `{type}`")
-	errNotRegistered       = errors.DefineNotFound("not_registered", "not registered")
-	errUnexpectedResult    = errors.Define("unexpected_result", "unexpected result code `{result_code}`", "result_description")
-	errCallerNotAuthorized = errors.DefinePermissionDenied("caller_not_authorized", "caller is not authorized for `{target}`")
-	errUnauthenticated     = errors.DefineUnauthenticated("unauthenticated", "unauthenticated")
-	errInvalidVendorID     = errors.DefineInvalidArgument("invalid_vendor_id", "invalid vendor ID")
+	errNoPublicTLSAddress = errors.DefineFailedPrecondition("no_public_tls_address",
+		"no public TLS address configured for interop",
+	)
+	errUnknownMACVersion  = errors.DefineInvalidArgument("unknown_mac_version", "unknown MAC version")
+	errInvalidLength      = errors.DefineInvalidArgument("invalid_length", "invalid length")
+	errInvalidRequestType = errors.DefineInvalidArgument("invalid_request_type", "invalid request type `{type}`")
+	errNotRegistered      = errors.DefineNotFound("not_registered", "not registered")
+	errUnexpectedResult   = errors.Define("unexpected_result",
+		"unexpected result code `{result_code}`", "result_description",
+	)
+	errCallerNotAuthorized = errors.DefinePermissionDenied("caller_not_authorized",
+		"caller is not authorized for `{target}`",
+	)
+	errUnauthenticated = errors.DefineUnauthenticated("unauthenticated", "unauthenticated")
+	errInvalidVendorID = errors.DefineInvalidArgument("invalid_vendor_id", "invalid vendor ID")
+)
 
+// LoRaWAN Backend Interfaces 1.0 errors.
+var (
 	ErrNoAction           = errors.DefineAborted("no_action", "no action", "result_description")
 	ErrMIC                = errors.DefineInvalidArgument("mic", "MIC failed", "result_description")
 	ErrFrameReplayed      = errors.DefineAborted("frame_replayed", "frame replayed", "result_description")
 	ErrJoinReq            = errors.DefineAborted("join_req", "join-request failed", "result_description")
-	ErrNoRoamingAgreement = errors.DefineFailedPrecondition("no_roaming_agreement", "no roaming agreement", "result_description")
-	ErrDeviceRoaming      = errors.DefineFailedPrecondition("device_roaming", "device roaming disallowed", "result_description")
-	ErrRoamingActivation  = errors.DefineFailedPrecondition("roaming_activation", "roaming activation disallowed", "result_description")
-	ErrActivation         = errors.DefineFailedPrecondition("activation", "activation disallowed", "result_description")
-	ErrUnknownDevEUI      = errors.DefineNotFound("unknown_dev_eui", "unknown DevEUI", "result_description")
-	ErrUnknownDevAddr     = errors.DefineNotFound("unknown_dev_addr", "unknown DevAddr", "result_description")
-	ErrUnknownSender      = errors.DefineNotFound("unknown_sender", "unknown sender", "result_description")
-	ErrUnknownReceiver    = errors.DefineNotFound("unknown_receiver", "unknown receiver", "result_description")
-	ErrDeferred           = errors.DefineAborted("deferred", "deferred", "result_description")
-	ErrTransmitFailed     = errors.DefineAborted("transmit_failed", "transmit failed", "result_description")
-	ErrFPort              = errors.DefineInvalidArgument("f_port", "invalid FPort", "result_description")
-	ErrProtocolVersion    = errors.DefineInvalidArgument("protocol_version", "invalid protocol version", "result_description")
-	ErrStaleDeviceProfile = errors.DefineFailedPrecondition("stale_device_profile", "stale device profile", "result_description")
-	ErrMalformedMessage   = errors.DefineInvalidArgument("malformed_message", "malformed message", "result_description")
-	ErrFrameSize          = errors.DefineInvalidArgument("frame_size", "frame size error", "result_description")
+	ErrNoRoamingAgreement = errors.DefineFailedPrecondition("no_roaming_agreement", "no roaming agreement",
+		"result_description",
+	)
+	ErrDeviceRoaming = errors.DefineFailedPrecondition("device_roaming", "device roaming disallowed",
+		"result_description",
+	)
+	ErrRoamingActivation = errors.DefineFailedPrecondition("roaming_activation", "roaming activation disallowed",
+		"result_description",
+	)
+	ErrActivation      = errors.DefineFailedPrecondition("activation", "activation disallowed", "result_description")
+	ErrUnknownDevEUI   = errors.DefineNotFound("unknown_dev_eui", "unknown DevEUI", "result_description")
+	ErrUnknownDevAddr  = errors.DefineNotFound("unknown_dev_addr", "unknown DevAddr", "result_description")
+	ErrUnknownSender   = errors.DefineNotFound("unknown_sender", "unknown sender", "result_description")
+	ErrUnknownReceiver = errors.DefineNotFound("unknown_receiver", "unknown receiver", "result_description")
+	ErrDeferred        = errors.DefineAborted("deferred", "deferred", "result_description")
+	ErrTransmitFailed  = errors.DefineAborted("transmit_failed", "transmit failed", "result_description")
+	ErrFPort           = errors.DefineInvalidArgument("f_port", "invalid FPort", "result_description")
+	ErrProtocolVersion = errors.DefineInvalidArgument("protocol_version", "invalid protocol version",
+		"result_description",
+	)
+	ErrStaleDeviceProfile = errors.DefineFailedPrecondition("stale_device_profile", "stale device profile",
+		"result_description",
+	)
+	ErrMalformedMessage = errors.DefineInvalidArgument("malformed_message", "malformed message", "result_description")
+	ErrFrameSize        = errors.DefineInvalidArgument("frame_size", "frame size error", "result_description")
 
 	resultErrors = map[ResultCode]*errors.Definition{
 		ResultNoAction:               ErrNoAction,