+++ pkg/interop/server_authn_mtls.go
@@ -19,8 +19,11 @@ import (
 	"crypto/tls"
 )
 
-func (s *Server) verifySenderCertificate(ctx context.Context, senderID string, state *tls.ConnectionState) (addrs []string, err error) {
-	// TODO: Support reading TLS client certificate from proxy headers (https://github.com/TheThingsNetwork/lorawan-stack/issues/717).
+func (s *Server) verifySenderCertificate(
+	ctx context.Context, senderID string, state *tls.ConnectionState,
+) (addrs []string, err error) {
+	// TODO: Support reading TLS client certificate from proxy header.
+	// (https://github.com/TheThingsNetwork/lorawan-stack/issues/717)
 	senderClientCAs, err := s.SenderClientCAs(ctx, senderID)
 	if err != nil {
 		return nil, err
@@ -29,7 +32,8 @@ func (s *Server) verifySenderCertificate(ctx context.Context, senderID string, s
 		peerCert, clientCA := chain[0], chain[len(chain)-1]
 		for _, senderClientCA := range senderClientCAs {
 			if clientCA.Equal(senderClientCA) {
-				// If the TLS client certificate contains DNS addresses, use those. Otherwise, fallback to using CommonName as address.
+				// If the TLS client certificate contains DNS addresses, use those.
+				// Otherwise, fallback to using CommonName as address.
 				if len(peerCert.DNSNames) > 0 {
 					addrs = append([]string(nil), peerCert.DNSNames...)
 				} else {
@@ -39,7 +43,7 @@ func (s *Server) verifySenderCertificate(ctx context.Context, senderID string, s
 			}
 		}
 	}
-	// TODO: Verify state.PeerCertificates[0] with senderClientCAs as Roots and state.PeerCertificates[1:] as Intermediates.
-	// (https://github.com/TheThingsNetwork/lorawan-stack/issues/718).
+	// TODO: Verify state.PeerCertificates[0] with senderClientCAs as Roots
+	// and state.PeerCertificates[1:] as Intermediates (https://github.com/TheThingsNetwork/lorawan-stack/issues/718).
 	return nil, errUnauthenticated.New()
 }